test(e2e): consolidate into one serial spec, fix route/selector bugs
- Merge user-registration-flow, offer-request-flow, and admin-management
  into a single core-flows.spec.ts using test.describe.serial. All spec
  files share one conductor/agent identity for the whole Playwright run
  (global-setup/teardown run once, not per file), and create_user throws
  UserAlreadyExists on a second call from the same agent, so independent
  files each trying to register their own user only works for the first
  file to execute.
- Fix wrong /profile route (actual route is /user).
- Fix home-page assertion that assumed a redirect to /profile/setup or
  /register; the home page instead shows a "Create Profile" CTA link
  with no navigation for agents with no profile.
- Fix offers-page empty-state assertion (text=No offers never matches;
  actual string is "No active offers found...").
- Simplify ui/package.json e2e scripts now that there is one spec file.
- Update documentation/guides/testing.md to match.
