+++ ui/package.json
@@ -10,15 +10,11 @@
     "test:unit": "vitest run tests/unit",
     "test:integration": "vitest run tests/integration",
     "test:e2e": "cross-env E2E_UI_PORT=8880 playwright test",
-    "test:e2e:holochain": "cross-env E2E_UI_PORT=8880 playwright test tests/e2e/specs",
     "test:e2e:ui": "cross-env E2E_UI_PORT=8880 playwright test --ui",
     "test:e2e:debug": "cross-env E2E_UI_PORT=8880 playwright test --debug",
-    "test:e2e:verbose": "cross-env E2E_UI_PORT=8880 E2E_VERBOSE=true playwright test tests/e2e/specs",
-    "test:e2e:ci": "cross-env CI=true E2E_UI_PORT=8880 playwright test tests/e2e/specs --reporter=html,junit,json",
-    "test:e2e:ci:all-browsers": "cross-env CI=true E2E_UI_PORT=8880 playwright test tests/e2e/specs --project=chromium-desktop --project=firefox --project=webkit",
-    "test:e2e:smoke": "cross-env E2E_UI_PORT=8880 playwright test tests/e2e/specs/user-registration-flow.spec.ts",
-    "test:e2e:offers": "cross-env E2E_UI_PORT=8880 playwright test tests/e2e/specs/offer-request-flow.spec.ts",
-    "test:e2e:admin": "cross-env E2E_UI_PORT=8880 playwright test tests/e2e/specs/admin-management.spec.ts",
+    "test:e2e:verbose": "cross-env E2E_UI_PORT=8880 E2E_VERBOSE=true playwright test",
+    "test:e2e:ci": "cross-env CI=true E2E_UI_PORT=8880 playwright test --reporter=html,junit,json",
+    "test:e2e:ci:all-browsers": "cross-env CI=true E2E_UI_PORT=8880 playwright test --project=chromium-desktop --project=firefox --project=webkit",
     "test:e2e:report": "playwright show-report test-results/html-report",
     "lint": "eslint .",
     "format": "prettier --write ."
